requirements.txt
# Core
streamlit>=1.32.0
numpy>=1.24.0
pandas>=1.5.3
python-dotenv>=1.0.0

# PDF processing
pdfplumber>=0.10.2
PyPDF2>=3.0.1
pdf2image>=1.16.3
Pillow>=9.4.0

# Langchain and Embeddings
langchain>=0.1.14
langchain-google-genai>=0.0.7
langchain-community

# Vector DB
faiss-cpu>=1.7.4
chromadb

# LLM APIs
google-generativeai>=0.3.2
groq>=0.4.2
